Added DescriptorProcessor which provides the logical implementation of the functions which the new DescriptorSystem relies upon.
This provides the following functionality:
- CreateDescriptor - Create a new EntityStateDescriptor on the target object. Uses a CreateDescriptorStateList to define the conditions
- GetDescriptor - Takes a target and the name of a descriptor to retrieve it. Considers conditional descriptors of the same identity before returning.

TryDescriptorFallback is a private function which works within GetDescriptor.
Its purpose is to modify the requested descriptor with any containing fallbacks that it knows about, to try to catch a missing descriptor.

This allows for more iterative design processes for new entities, instead of requiring every descriptor to be made, all at once.

- GenerateAdditiveDescriptors - These new descriptors take prefix, infix and suffix from the object, and use them to assemble a large, additive description from the contents of the subjects ItemContainer inventory.

The use case for this is to present information about containers.

For example, items on a table, or items on a shelf, or inside a glass cabinet.