New NumberedConsoleMenu completely takes over the main functionality of the former ExploreMenu.
The main logical loop of the menu is stored here in RunMenu.
Each logical process of assessing, drawing, recieving input and redrawing is done here.

Virtuals have been exposed to override this functionality, but base implementations which should be replicated by each menu are put here.

Menus can override every single element which impacts their visual representation, such as the order of interactions, the order of elements (the base of which should be changd with the UpdatePriority) and so on.